Issue: 1591
Protect against scripts being run on the wrong host.  DAQLaunch/Config/Status, DeployPDAQ, and
ExpControlSkel will all now check to see what hose you are running them on before continuing.  Which
host to allow is as per the table in the "PDAQ Command Quick Reference" section of the daq manual.

If the host name detection code is not working right or you REALLY want to run that script on a host
you should not be using you can use the -m or --no-host-check option.

The update to __init__.py is to check for ternary operator support.  Ie python prior to 2.5 doesn't
support them so check for it, otherwise you'll get a syntax error at runtime.